requirements.txt
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
configparser==7.1.0
confluent-kafka==2.5.3
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
minio==7.2.9
numpy==2.1.1
orjson==3.10.7
pandas==2.2.3
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
quixstreams==2.11.0
referencing==0.35.1
requests==2.32.3
rocksdict==0.3.23
rpds-py==0.20.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
